# backend/report-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: "Manages report templates, execution, history, and scheduling. Uses\
    \ PostgreSQL for relations and Elasticsearch for analytic report data."
  title: Report Service
  version: 1.0.0
servers:
- url: http://localhost:8083/api/v1/reports
...
paths:
...
  /execute/{id}:
    post:
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          description: Report execution triggered
      security:
      - bearerAuth: []
      summary: Execute a report template
      tags:
      - Reports
      x-accepts:
      - application/json
      x-tags:
      - tag: Reports
...
components:
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http